--- app/src/main/java/com/mvckx/elistique/ui/MainActivity.kt
@@ -1,13 +1,0 @@
-package com.mvckx.elistique.ui
-
-import androidx.appcompat.app.AppCompatActivity
-import android.os.Bundle
-import com.mvckx.elistique.R
-
-class MainActivity : AppCompatActivity() {
-
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        setContentView(R.layout.activity_main)
-    }
-}
