+++ app/src/main/res/layout/activity_categories.xml
@@ -3,10 +3,11 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ui.MainActivity">
-    <fragment
-        android:id="@+id/fragment"
-        android:name="com.mvckx.elistique.ui.categories.CategoriesFragment"
+    tools:context=".ui.categories.CategoriesActivity">
+
+
+    <androidx.recyclerview.widget.RecyclerView
+        android:id="@+id/recyclerView"
         android:layout_width="match_parent"
-        android:layout_height="match_parent"/>
+        android:layout_height="match_parent" />
 </FrameLayout>
